upgrade steps, dates
