--- pages/_app.js
@@ -1,8 +1,0 @@
-import '../styles/globals.css'
-import '../node_modules/toastr/build/toastr.css'
-
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
